+++ src/main/java/at/hannibal2/skyhanni/events/fishing/SeaCreatureFishEvent.kt
@@ -1,6 +1,7 @@
-package at.hannibal2.skyhanni.events
+package at.hannibal2.skyhanni.events.fishing
 
 import at.hannibal2.skyhanni.api.event.SkyHanniEvent
+import at.hannibal2.skyhanni.events.LorenzChatEvent
 import at.hannibal2.skyhanni.features.fishing.SeaCreature
 
 class SeaCreatureFishEvent(
